Store `gridsteps_inv` in `Grid` (#147)
* Store gridsteps_inv in Grid

* Use `Float64` for `Grid` by default

+++ test/grid.jl
@@ -1,17 +1,17 @@
 @testset "Grid" begin
     # constructors
-    @test @inferred(Grid(0:10))::Grid{1, Int} == Vec.(0:10)
-    @test @inferred(Grid(0:10, 0:20))::Grid{2, Int} == Vec.(collect(Iterators.product(0:10, 0:20)))
-    @test @inferred(Grid(0:10, 0:20, 0:30))::Grid{3, Int} == Vec.(collect(Iterators.product(0:10, 0:20, 0:30)))
-    @test @inferred(Grid(LinearBSpline(), 0:10))::Grid{1, Int} == Grid(0:10)
-    @test @inferred(Grid(LinearBSpline(), 0:10, 0:20))::Grid{2, Int} == Grid(0:10, 0:20)
-    @test @inferred(Grid(LinearBSpline(), 0:10, 0:20, 0:30))::Grid{3, Int} == Grid(0:10, 0:20, 0:30)
+    @test @inferred(Grid(0:10))::Grid{1, Float64} == Vec.(0:10)
+    @test @inferred(Grid(0:10, 0:20))::Grid{2, Float64} == Vec.(collect(Iterators.product(0:10, 0:20)))
+    @test @inferred(Grid(0:10, 0:20, 0:30))::Grid{3, Float64} == Vec.(collect(Iterators.product(0:10, 0:20, 0:30)))
+    @test @inferred(Grid(LinearBSpline(), 0:10))::Grid{1, Float64} == Grid(0:10)
+    @test @inferred(Grid(LinearBSpline(), 0:10, 0:20))::Grid{2, Float64} == Grid(0:10, 0:20)
+    @test @inferred(Grid(LinearBSpline(), 0:10, 0:20, 0:30))::Grid{3, Float64} == Grid(0:10, 0:20, 0:30)
     @test_throws MethodError Grid(NodeState, 0:10)
     @test_throws MethodError Grid(NodeState, 0:10, 0:20)
     @test_throws MethodError Grid(NodeState, 0:10, 0:20, 0:30)
-    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10))::Grid{1, Int} == Grid(0:10)
-    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10, 0:20))::Grid{2, Int} == Grid(0:10, 0:20)
-    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10, 0:20, 0:30))::Grid{3, Int} == Grid(0:10, 0:20, 0:30)
+    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10))::Grid{1, Float64} == Grid(0:10)
+    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10, 0:20))::Grid{2, Float64} == Grid(0:10, 0:20)
+    @test @inferred(Grid(NodeState, LinearWLS(LinearBSpline()), 0:10, 0:20, 0:30))::Grid{3, Float64} == Grid(0:10, 0:20, 0:30)
 
     # gridsteps/gridaxes/gridorigin
     grid = Grid(CubicBSpline(), 0:1.0:10, 1:2.0:20)
